# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Penaup Runtime API
  version: 0.1.0
  description: |
    花生片 Penaup 的用户、设备、媒体和显影状态接口。
    照片本体不经过 MQTT；Web、微信小程序和未来 iOS 使用同一套 transfer 状态。
  license:
    name: Poboll Non-Commercial License v1.0
    identifier: LicenseRef-Poboll-NonCommercial
servers:
  - url: https://penaup.example.com
    description: 生产部署示例
  - url: http://127.0.0.1:8787
    description: 本地运行时
...
paths:
...
  /api/v1/events/stream:
    get:
      tags: [events]
      security:
        - BearerAuth: []
        - CookieSession: []
      summary: SSE 实时事件
      description: iOS 可用 URLSession bytes/AsyncStream 消费；断线后使用 transfer 查询接口补齐状态。
      responses:
        '200':
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
                description: event 行包含 type 与 JSON data
        '401': { description: 未登录 }
...
  /api/v1/device/heartbeat:
    get:
      tags: [device]
      security: []
      summary: 旧固件 HTTP 心跳与命令领取
      parameters:
        - $ref: '#/components/parameters/DeviceIdQuery'
        - $ref: '#/components/parameters/DeviceTokenQuery'
        - { $ref: '#/components/parameters/DeviceModelQuery' }
        - { $ref: '#/components/parameters/BatteryQuery' }
        - { $ref: '#/components/parameters/DeviceStateQuery' }
      responses:
        '200': { description: 心跳成功并返回待执行命令 }
        '400': { description: 缺少 device_id }
        '401': { description: 设备 token 无效 }
  /api/v1/device/status:
    get:
      tags: [device]
      security: []
      summary: 查询设备实时状态
      parameters:
        - $ref: '#/components/parameters/DeviceIdQuery'
        - $ref: '#/components/parameters/DeviceTokenQuery'
      responses:
        '200': { description: 设备状态 }
        '401': { description: 设备 token 无效 }
  /api/v1/device/film/latest.film:
    get:
      tags: [device]
      security: []
      summary: 下载设备所属最新 film
      parameters:
        - $ref: '#/components/parameters/DeviceIdQuery'
        - $ref: '#/components/parameters/DeviceTokenQuery'
      responses:
        '200': { description: film 文件流 }
        '401': { description: 设备 token 无效 }
        '404': { description: film 不存在 }
  /api/v1/device/film/{filename}:
    get:
      tags: [device]
      security: []
      summary: 按文件名下载设备可访问的 film
      parameters:
        - name: filename
          in: path
          required: true
          schema: { type: string, maxLength: 96 }
        - $ref: '#/components/parameters/DeviceIdQuery'
        - $ref: '#/components/parameters/DeviceTokenQuery'
      responses:
        '200': { description: film 文件流 }
        '401': { description: 设备 token 无效 }
        '404': { description: film 不存在 }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    CookieSession:
      type: apiKey
      in: cookie
      name: penaup_session
    WereadSkillKey:
      type: apiKey
      in: header
      name: X-Penaup-WeRead-Key
  parameters:
...
    DeviceIdQuery:
      name: device_id
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    DeviceTokenQuery:
      name: token
      in: query
      required: false
      schema: { type: string, maxLength: 256 }
...
    DeviceModelQuery:
      name: model
      in: query
      required: false
      schema: { type: string, enum: [PENAUP_STD, PENAUP_PRO, PENAUP_MAX] }
    BatteryQuery:
      name: battery
      in: query
      required: false
      schema: { type: number, minimum: -1, maximum: 100 }
    DeviceStateQuery:
      name: state
      in: query
      required: false
      schema: { type: string, enum: [idle, preparing, transferring, refreshing, succeeded, failed, unknown] }
...